backend/requirements.txt
fastapi
uvicorn[standard]
pyyaml
networkx
numpy
scipy
tabulate
pytest
pydantic
reportlab
